Added verify email code validation when registering a user

+++ migrations/20191007062416-create-users.js
@@ -45,7 +45,7 @@ module.exports = {
                 type: Sequelize.STRING
             },
             verification_code : {
-                type: Sequelize.TINYINT
+                type: Sequelize.INTEGER
             },
             created_at: {
                 defaultValue: Sequelize.fn('NOW'),
